feat: Enhance waste tracking functionality with barcode scanning and data fetching
- Implemented barcode scanning in WasteDetailPage with error handling and loading states.
- Integrated API calls to fetch waste data based on scanned barcode.
- Added confirmation dialog for saving waste records with user authentication.
- Improved user experience with loading indicators and error messages.
- Updated WasteSortingPage to fetch waste materials and categories from the API.
- Introduced pagination for waste materials display and search functionality.
- Created new interfaces for Waste and History items to standardize data handling.
- Enhanced UI components for better visual feedback during data operations.

+++ backend/src/waste/entities/waste.entity.ts
@@ -24,7 +24,7 @@ export class Waste {
   @Column({ type: 'varchar', length: 255, nullable: true })
   waste_image: string;
 
-  @Column({ type: 'int', nullable: true })
+  @Column({ type: 'bigint', nullable: true })
   barcode: number;
 
   @CreateDateColumn({ type: 'date' })
